+++ packages/react-vapor/src/components/drop/DropPod.tsx
@@ -53,8 +53,6 @@ class RDropPod extends React.PureComponent<IRDropPodProps, IDropPodState> {
         minWidth: 0,
         minHeight: 0,
         hasSameWidth: false,
-        selector: Defaults.DROP_ROOT,
-        parentSelector: Defaults.DROP_PARENT_ROOT,
     };
     private parentMutationObserver: MutationObserver;
 
@@ -134,16 +132,21 @@ class RDropPod extends React.PureComponent<IRDropPodProps, IDropPodState> {
         return this.dropRef.current && this.props.isOpen && !!this.props.positions.length;
     }
 
+    private getRelativeParent = () => {
+        const closestCurrentRef = this.props.buttonRef.current?.closest(
+            this.props.parentSelector ?? Defaults.DROP_PARENT_ROOT
+        );
+        return closestCurrentRef ?? this.props.buttonRef.current?.parentElement;
+    };
+
     private calculateStyleOffset(): React.CSSProperties {
         let newDomPosition: IDomPositionCalculatorReturn = {};
         if (this.canRenderDrop()) {
             const buttonOffset: ClientRect | DOMRect =
                 this.props.buttonRef.current?.getBoundingClientRect() ?? this.state.offset;
             const dropOffset: ClientRect | DOMRect = this.dropRef.current.getBoundingClientRect();
-            const relativeParent =
-                this.props.buttonRef.current?.closest(this.props.parentSelector) ??
-                this.props.buttonRef.current?.parentElement;
 
+            const relativeParent = this.getRelativeParent();
             const parentOffset = relativeParent.getBoundingClientRect();
 
             const boundingLimit: IBoundingLimit = {
@@ -234,7 +237,7 @@ class RDropPod extends React.PureComponent<IRDropPodProps, IDropPodState> {
             this.lastPosition
         );
 
-        return ReactDOM.createPortal(drop, document.querySelector(this.props.selector));
+        return ReactDOM.createPortal(drop, document.querySelector(this.props.selector ?? Defaults.DROP_ROOT));
     }
 }